CloudFunction/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-pubsub==2.10.0